% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann2term.R
\name{ann2term}
\alias{ann2term}
\title{Annotate pathways}
\usage{
ann2term(MScores)
}
\arguments{
\item{MScores}{Matrix with pathways IDs as row names}
}
\value{
A dataframe with the input IDs and their corresponding annotations
}
\description{
Annotate pathways
}
\examples{
data(refData, exampleData, exampleMetadata)
\donttest{
exampleRefMScore <- mScores_createReference(data=refData, genesets="tmod")
relevantPaths <- diseasePaths(MRef=exampleRefMScore,
min_datasets=3,
perc_samples=10)

MScoresExample <- getMscores(genesets = relevantPaths,
Patient = exampleData,
Healthy = NULL,
nk = 5)
}
MScoresAnnotated <- ann2term(MScoresExample)
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{mScores_createReference}}
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
