% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissectDB.R
\name{dissectDB}
\alias{dissectDB}
\title{Split pathways into coexpressed subpathways}
\usage{
dissectDB(
  refData,
  geneSets,
  minPathSize = 10,
  minSplitSize = 3,
  maxSplits = NULL,
  explainedVariance = 60,
  percSharedGenes = 90
)
}
\arguments{
\item{refData}{A refData object structure: a list of lists, each one with a
cases expression matrix and controls expression matrix
(named as Disease and Healthy). It can be constructed with the buildRefObject
 function.}

\item{geneSets}{A named list with each
gene set or the name of one preloaded database (gobp, gomf, gocc,
kegg, reactome, tmod).}

\item{minPathSize}{numeric, minimum number of genes in a pathway to consider
splitting it.}

\item{minSplitSize}{numeric, minimum number of genes in a subpathway.
Smaller splits will be merged with the closest coexpressed subpathway.}

\item{maxSplits}{numeric, maximum number of subpathways for a pathway. If
NULL (default), there is not limit.}

\item{explainedVariance}{numeric, percentage of cumulative variance explained
within a pathway. This parameter is used to select the number of subdivisions
of a pathway that manage to explain at least the percentage of variance
defined by explainedVariance.}

\item{percSharedGenes}{numeric, minimum percentage of common genes across
datasets to merge them before clustering. If NULL or this percentage is not
reached, clustering is performed for each dataset independently and
consensus subpathways are obtained from co-occurrence across datasets.}
}
\value{
A list with the subpathways.
}
\description{
Split pathways into coexpressed subpathways
}
\examples{
data(refData)
set.seed(1234)
custom.tmod <- dissectDB(refData, geneSets="tmod")
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{buildRefObject}}, \code{\link{mScores_createReference}},
 \code{\link{getScores}}
}
\author{
Jordi Martorell-Marugán, \email{jordi.martorell@genyo.es}

Daniel Toro-Dominguez, \email{danieltorodominguez@gmail.com}
}
