% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRefMScore-data.R
\docType{data}
\name{exampleRefMScore}
\alias{exampleRefMScore}
\title{Example of reference M-scores}
\format{
An object of class \code{"list"} with...
}
\usage{
data(exampleRefMScore)
}
\description{
Output of the getMscoresRef() function with refData
}
\keyword{datasets}
