% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetsData-data.R
\docType{data}
\name{genesetsData}
\alias{genesetsData}
\title{Preloaded gene sets}
\format{
An object of class \code{"list"} with one list per database. Each 
database consists on a list of gene sets, containing the gene symbols 
associated to it.
}
\usage{
data(genesetsData)
}
\description{
genesetsData was constructed from the GeneCodis database 
(https://genecodis.genyo.es/)
}
\keyword{datasets}
