% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getML.R
\name{getML}
\alias{getML}
\title{Build machine-learning based on gene-expression/mscores data}
\usage{
getML(
  expData,
  metadata,
  algorithms = "all",
  add = NULL,
  var2predict,
  subsamples = 4,
  splitProp = 0.75,
  foldsCV = 10,
  repeatsCV = 5,
  positiveClass = NULL,
  featureFilter = "none",
  prior = "AUC"
)
}
\arguments{
\item{expData}{Feature matrix. Samples in columns and features in
rows. expData must be a numerical matrix}

\item{metadata}{dataframe with information for each sample. Samples in rows
and variables in columns}

\item{algorithms}{'glm','lm','lda','xgbTree','rf','knn','svmLinear','nnet',
'svmRadial','nb','lars','rpart', 'gamboost', 'ada', 'brnn', 'enet' or 'all'
(all algorithms are used)}

\item{add}{additional model list (supported by caret) in format:
add=list(modelName = caretModelSpec(method='modelName',
tuneGrid=(.parameters='values')))}

\item{var2predict}{character with the column name of the @metadata to predict}

\item{subsamples}{number of sub-samples in which exp.data is divided}

\item{splitProp}{proportion of samples used as train for the subsamples (0-1)}

\item{foldsCV}{number of internal folds (for parameter tuning)}

\item{repeatsCV}{number of repetitions of the parameter tuning process}

\item{positiveClass}{outcome vaue that must be considered as positive class
(for categoric outcomes)}

\item{featureFilter}{method to reduce number of features (none,fcbf)}

\item{prior}{rank best model based on AUC (AUC), the mean of AUC and the
balanced accuracy (BAUC), the balanced accuracy (BA) or manual (Manual)}
}
\value{
A list with four elements. The first one is the model. The second one
is a table with different metrics obtained. The third one is a list with the
best parameters selected in tuning process. The last element contains data
for AUC plots
}
\description{
Retrieve best model for a specific outcome
}
\examples{
data(refData, exampleData, exampleMetadata)
\donttest{
exampleRefMScore <- getMscoresRef(data=refData, genesets="tmod")
relevantPaths <- diseasePaths(MRef=exampleRefMScore,
min_datasets=3,
perc_samples=10)

MScoresExample <- getMscores(genesets = relevantPaths,
Patient = exampleData,
Healthy = NULL,
nk = 5)

fit.model <- getML(expData=MScoresExample,
metadata=exampleMetadata,
var2predict="Response",
algorithms="svmLinear",
subsamples=5,
splitProp=0.8,
foldsCV=10,
repeatsCV=10,
featureFilter = "none",
prior="BAUC")
}

}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
