% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMscores.R
\name{getMscores}
\alias{getMscores}
\title{Calculate M-scores for a dataset}
\usage{
getMscores(Patient, Healthy = NULL, genesets, nk = 5, cores = 1)
}
\arguments{
\item{Patient}{Expression matrix of cases or numeric vector with one sample.}

\item{Healthy}{Expression matrix of healthy controls.}

\item{genesets}{Output from the diseasePaths function}

\item{nk}{Only for Healthy=NULL, Number of most similar samples to impute
M-scores.}

\item{cores}{Number of cores to be used.}
}
\value{
A list with the results of each of the analyzed regions. For each
region type, a data frame with the results and a list with the probes
associated to each region are generated. In addition, this list also contains
the input methData, pheno and platform objects
}
\description{
Subtitle
}
\examples{
data(refData, exampleData)
exampleRefMScore <- getMscoresRef(data=refData, genesets="tmod")
relevantPaths <- diseasePaths(MRef=exampleRefMScore, min_datasets=3,
perc_samples=10)
MScoresExample <- GetMscores(genesets = relevantPaths, Patient = exampleData,
Healthy = NULL, nk = 5)
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{diseasePaths}}, \code{\link{getML}}
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
