% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMscoresRef.R
\name{getMscoresRef}
\alias{getMscoresRef}
\title{Calculate M-scores from reference data}
\usage{
getMscoresRef(data, genesets = "reactome", customGeneset = NULL, cores = 1)
}
\arguments{
\item{data}{A list of lists, each one with a cases expression matrix and
controls expression matrix, always in this order.}

\item{genesets}{character, name of the preloaded database (gobp, gomf, gocc,
kegg, reactome, tmod) or 'custom' to provide the annotation.}

\item{customGeneset}{Only if genesets == 'custom'. A named list with each
gene set.}

\item{cores}{Number of cores to be used.}
}
\value{
A list with three elements. The first one is a list with the M-scores
for each dataset. The second one is the geneset used for the analysis and
the third one is the input gene expression data.
}
\description{
Retrieve Mscore for a list of datasets
}
\examples{
data(refData)
\donttest{
refMscore <- getMscoresRef(data=refData, genesets="tmod")
}
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{diseasePaths}}
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
