% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mScores_createReference.R
\name{mScores_createReference}
\alias{mScores_createReference}
\title{Create a reference dataset based on M-scores}
\usage{
mScores_createReference(refData, geneSets, cores = 1)
}
\arguments{
\item{refData}{A refData object structure: a list of lists, each one with a
cases expression matrix and controls expression matrix
(named as Disease and Healthy). It can be constructed with the buildRefObject
 function.}

\item{geneSets}{A named list with each
gene set or the name of one preloaded database (gobp, gomf, gocc,
kegg, reactome, tmod)}

\item{cores}{Number of cores to be used.}
}
\value{
A list with three elements. The first one is a list with the M-scores
for each dataset. The second one is the geneSet used for the analysis and
the third one is the input data.
}
\description{
Create a reference dataset based on M-scores
}
\examples{
data(refData)
refMscore <- mScores_createReference(refData, geneSets="tmod")
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{mScores_imputeFromReference}}, \code{\link{dissectDB}},
\code{\link{mScores_filterPaths}}, \code{\link{trainModel}}
}
\author{
Jordi Martorell-Marugán, \email{jordi.martorell@genyo.es}

Daniel Toro-Dominguez, \email{danieltorodominguez@gmail.com}
}
