% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathMEDExampleData-data.R
\docType{data}
\name{pathMEDExampleData}
\alias{pathMEDExampleData}
\title{Example of test gene expression data}
\format{
An object of class \code{"data.frame"} with genes in rows and samples
in columns.
}
\usage{
data(pathMEDExampleData)
}
\description{
pathMEDExampleData was obtained from a dataset downloaded from NCBI GEO 
(GSE224705),
that contains lupus patients treated with Micophenolate mofetil. The same
preprocessing was done as for the datasets used to create refData. 40
patients were randomly selected, 20 samples from responding patients and 20
from non-responders.
}
\keyword{datasets}
