% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathMEDExampleMetadata-data.R
\docType{data}
\name{pathMEDExampleMetadata}
\alias{pathMEDExampleMetadata}
\title{Metadata of test gene expression data}
\format{
An object of class \code{"data.frame"} with samples in rows and
variables in columns.
}
\usage{
data(pathMEDExampleMetadata)
}
\description{
Metadata from the dataset GSE224705. Response column conteins the information
about the response and non-response to the drug for each sample.
}
\keyword{datasets}
