% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRefData.R
\name{buildRefData}
\alias{buildRefData}
\title{Establish reference data structure for input to the pathMED pipeline}
\usage{
buildRefData(data, metadata, groupVar, healthyGroup)
}
\arguments{
\item{data}{A list of data frames or a single data frame with samples in 
columns and features in rows.}

\item{metadata}{A list of data frames or a single data frame with information
for each sample. Samples in rows and variables in columns.}

\item{groupVar}{Character or list of characters indicating the column name of
@metadata classifying the samples in healthy and disease. If several metadatas
 are provided a @groupVar can be specified for each metadata.}

\item{healthyGroup}{Character or list of characters indicating which @groupVar
level corresponds to healthy samples. All other samples will be 
considered as disease. If several @groupVar are provided a @healthyGroup
can be specified for each @groupVar}
}
\value{
A list of reference Healthy and Disease data frames that serves as 
input for getMscoresRef and splitPathways functions.
}
\description{
Establish reference data structure for input to the pathMED pipeline
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{buildRefData}}
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
