% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRefObject.R
\name{buildRefObject}
\alias{buildRefObject}
\title{Create a reference data object for input to the pathMED functions}
\usage{
buildRefObject(data, metadata, groupVar, controlGroup)
}
\arguments{
\item{data}{A list of data frames or a single data frame with samples in
columns and features in rows.}

\item{metadata}{A list of data frames or a single data frame with information
for each sample. Samples in rows and variables in columns.}

\item{groupVar}{Character or list of characters indicating the column name of
@metadata classifying the samples in controls and cases. If several metadata
objects are provided a @groupVar can be specified for each metadata.}

\item{controlGroup}{Character or list of characters indicating which @groupVar
level corresponds to the control group, usually healthy samples. All other
samples will be considered as cases, usually disease samples. If
several @groupVar are provided a @controlGroup can be specified
for each @groupVar}
}
\value{
A refData object that serves as input
 for mScores_createReference and dissectDB functions.
}
\description{
Create a reference data object for input to the pathMED functions
}
\examples{
data(reference_datasets)

refData <- buildRefObject(
  data = list(dataset1, dataset2, dataset3, dataset4),
  metadata = list(metadata1, metadata2, metadata3, metadata4),
  groupVar = "group",
  controlGroup = "Healthy_sample")

## Also works with a metadata for all datasets
metadata <- rbind(metadata1, metadata2, metadata3, metadata4)
refData <- buildRefObject(
  data = list(dataset1, dataset2, dataset3, dataset4),
  metadata = metadata,
  groupVar = "group",
  controlGroup = "Healthy_sample")

}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{mScores_createReference}}, \code{\link{dissectDB}}
}
\author{
Iván Ellson, \email{ivan.ellson.l@gmail.com }

Jordi Martorell-Marugán, \email{jordi.martorell@genyo.es}

Daniel Toro-Dominguez, \email{danieltorodominguez@gmail.com}
}
