% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasePaths.R
\name{diseasePaths}
\alias{diseasePaths}
\title{Find relevant pathways from the reference M-scores}
\usage{
diseasePaths(
  MRef,
  min_datasets = round(length(MRef[[1]]) * 0.34),
  perc_samples = 10,
  Pcutoff = 0.05
)
}
\arguments{
\item{MRef}{output from the getMscoresRef function}

\item{min_datasets}{number of datasets that each pathway must meet the
perc_samples threshold}

\item{perc_samples:}{minimun percentage of samples in a dataset in which a
pathway must be significant}

\item{Pcutoff:}{P-value cutoff for significance}
}
\value{
A list with the selected pathways
}
\description{
Subtitle
}
\examples{
data(pathMED)
\donttest{
refMscore <- getMscoresRef(data=refData, genesets="tmod")
}
relevantPaths <- diseasePaths(MRef=refMscore, min_datasets=3,
perc_samples=10)
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{getMscoresRef}}, \code{\link{getMscores}}
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
