% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissectDB.R
\name{dissectDB}
\alias{dissectDB}
\title{Cluster pathways into pathway-components}
\usage{
dissectDB(
  data,
  genesets,
  customGeneset = NULL,
  minPathSize = 10,
  minSplitSize = 3,
  maxSplits = NULL,
  explainedvariance = 60,
  percSharedgenes = 90
)
}
\arguments{
\item{data}{A refData object structure: a list of lists, each one with a
cases expression matrix and controls expression matrix (named as Disease and Healthy)}

\item{genesets}{character, name of the preloaded database (gobp, gomf, gocc,
kegg, reactome, tmod) or 'custom' to provide the annotation.}

\item{customGeneset}{Only if genesets == 'custom'. A named list with each
gene set.}

\item{minPathSize}{numeric, minimum number of genes in a pathway to consider
split the pathway}

\item{minSplitSize}{numeric, minimum number of genes in a pathway-split.
Smaller splits will join the closest split.}

\item{maxSplits}{numeric, maximum number of subdivisions for a pathway. NULL,
any restrictions}

\item{explainedvariance}{numeric, percentage of cumulative variance explained
within a pathway. This parameter is used to select the number of subdivisions
of a pathway that manage to explain at least the percentage of variance
defined by explainedvariance.}
}
\value{
A list with all pathways and pathway-subdivisions and their involved
genes
}
\description{
Cluster pathways into pathway-components
}
\examples{
data(refData)
set.seed(1234)
custom.tmod <- dissectDB(data=refData,
genesets="tmod",
customGeneset = NULL,
minPathSize = 10,
minSplitSize = 3,
maxSplits = 3,
explainedvariance = 50)
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{diseasePaths}}, \code{\link{getML}}
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
