% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData-data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example of test gene expression data}
\format{
An object of class \code{"data.frame"} with genes in rows and samples
in columns.
}
\usage{
data(exampleData)
}
\description{
Gene expression data from...
}
\keyword{datasets}
