% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleMetadata-data.R
\docType{data}
\name{exampleMetadata}
\alias{exampleMetadata}
\title{Metadata of test gene expression data}
\format{
An object of class \code{"data.frame"} with samples in rows and
variables in columns.
}
\usage{
data(exampleMetadata)
}
\description{
Metadata from the study...
}
\keyword{datasets}
