% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getML.R
\name{getML}
\alias{getML}
\title{Build machine-learning based on gene-expression/mscores data}
\usage{
getML(
  expData,
  metadata,
  models = methodsML(),
  var2predict,
  subsamples = 4,
  splitProp = 0.75,
  foldsCV = 10,
  repeatsCV = 5,
  positiveClass = NULL
)
}
\arguments{
\item{expData}{Feature matrix. Samples in columns and features in
rows. expData must be a numerical matrix}

\item{metadata}{dataframe with information for each sample. Samples in rows
and variables in columns}

\item{models}{named list with the ML models generated with
caret::caretModelSpec function. methodsML function may be used to prepare
this list}

\item{var2predict}{character with the column name of the @metadata to predict}

\item{subsamples}{number of sub-samples in which exp.data is split}

\item{splitProp}{proportion of samples used as train for the subsamples}

\item{foldsCV}{number of cross-validation folds (for parameter tuning)}

\item{repeatsCV}{number of repetitions of the parameter tuning process}

\item{positiveClass}{outcome value that must be considered as positive class
(for categoric outcomes)}
}
\value{
A list with four elements. The first one is the model. The second one
is a table with different metrics obtained. The third one is a list with the
best parameters selected in tuning process. The last element contains data
for AUC plots
}
\description{
Build machine-learning based on gene-expression/mscores data
}
\examples{
data(refData, exampleData, exampleMetadata)
\donttest{
exampleRefMScore <- getMscoresRef(data=refData, genesets="tmod")
relevantPaths <- diseasePaths(MRef=exampleRefMScore,
min_datasets=3,
perc_samples=10)

MScoresExample <- getMscores(genesets = relevantPaths,
Patient = exampleData,
Healthy = NULL,
nk = 5)

fit.model <- getML(expData=MScoresExample,
metadata=exampleMetadata,
var2predict="Response",
models=methodsML("svmLinear"),
subsamples=5,
splitProp=0.8,
foldsCV=10,
repeatsCV=10)
}

}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
