% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsML.R
\name{methodsML}
\alias{methodsML}
\title{Prepare ML models}
\usage{
methodsML(
  algorithms = c("glm", "lm", "lda", "xgbTree", "rf", "knn", "svmLinear", "svmRadial",
    "nnet", "nb", "lars", "rpart", "ada", "gamboost"),
  outcomeClass = NULL,
  tuneLength = 15
)
}
\arguments{
\item{algorithms}{'glm','lm','lda','xgbTree','rf','knn','svmLinear','nnet',
'svmRadial','nb','lars','rpart', 'gamboost', 'ada', 'brnn', 'enet' or 'all'
(all algorithms are used)}

\item{outcomeClass}{Predicted variable type ('character' or 'numeric')}

\item{tuneLength}{maximum number of tuning parameter combinations}
}
\value{
A list with the selected models ready to use as the 'models'
parameter in the getml() function
}
\description{
Prepare ML models
}
\examples{
methodsML(c('rf', 'knn'), tuneLength=20)
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{getML}}
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
