% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mScores_createReference.R
\name{mScores_createReference}
\alias{mScores_createReference}
\title{Create a reference of M-Scores}
\usage{
mScores_createReference(datasetsList, geneSets = "reactome", cores = 1)
}
\arguments{
\item{datasetsList}{A list of lists, each one with a cases data matrix and
controls data matrix, always in this order.}

\item{geneSets}{A named list with each
gene set or the name of one preloaded database (gobp, gomf, gocc,
kegg, reactome, tmod)}

\item{cores}{Number of cores to be used.}
}
\value{
A list with three elements. The first one is a list with the M-scores
for each dataset. The second one is the geneset used for the analysis and
the third one is the input data.
}
\description{
Create a reference of M-Scores
}
\examples{
data(refData)
refMscore <- mScores_createReference(datasetsList=refData, geneSets="tmod")
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{diseasePaths}}
}
\author{
Jordi Martorell-Marugán, \email{jordi.martorell@genyo.es}

Daniel Toro-Dominguez, \email{danieltorodominguez@gmail.com}
}
