% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mScores_imputeFromReference.R
\name{mScores_imputeFromReference}
\alias{mScores_imputeFromReference}
\title{Calculate M-Scores without healthy controls}
\usage{
mScores_imputeFromReference(
  inputData,
  geneSets,
  externalReference,
  nk = 25,
  imputeAll = FALSE,
  cores = 1
)
}
\arguments{
\item{inputData}{Data matrix or data frame.}

\item{geneSets}{A named list with each
gene set or the name of one preloaded database (gobp, gomf, gocc,
kegg, reactome, tmod).}

\item{externalReference}{External reference created with
the mScores_createReference function.}

\item{nk}{Number of
most similar samples from the external reference to impute M-scores.}

\item{imputeAll}{By default, only samples that do not
surpass the distance of 30 with the external reference are imputed. If TRUE,
impute all samples.}

\item{cores}{Number of cores to be used.}
}
\value{
A list with the results of each of the analyzed regions. For each
region type, a data frame with the results and a list with the probes
associated to each region are generated. In addition, this list also contains
the input methData, pheno and platform objects
}
\description{
Calculate M-Scores without healthy controls
}
\examples{
data(exampleData, refData)
MScoresExample <- mScores_imputeFromReference(exampleData, geneSets = "tmod",
                            externalReference=refData)
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\seealso{
\code{\link{mScores_filterPaths}}, \code{\link{trainModel}}
}
\author{
Daniel Toro-Dominguez, \email{daniel.toro@genyo.es}

Jordi Martorell-Marugan, \email{jordi.martorell@genyo.es}
}
