% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictExternal.R
\name{predictExternal}
\alias{predictExternal}
\title{Predict conditions in external datasets}
\usage{
predictExternal(testData, model, realValues = NULL, positiveClass = NULL)
}
\arguments{
\item{testData}{Numerical matrix with the same features used
for the model construction in rows, and the samples (new observations) in
columns}

\item{model}{trainModel output or a caret-like model object}

\item{realValues}{Optional, named vector (for numerical variables) or
named factor (for categorical variables) with real values for each sample}

\item{positiveClass}{Optional, positive class to get confusion matrix.
Only needed when realValues = TRUE and for categorical variables}
}
\value{
A dataframe with predictions (if realValues
is not provided) or a list with the dataframe with predictions and a
dataframe with the performance metrics (if realValues is provided)
}
\description{
Predict conditions in external datasets
}
\examples{
data(refData)

commonGenes <- intersect(rownames(dataset1), rownames(dataset2))
dataset1 <- dataset1[commonGenes, ]
dataset2 <- dataset2[commonGenes, ]

scoresExample <- getScores(dataset1, geneSets = "tmod", method = "Z-score")

set.seed(123)
trainedModel <- trainModel(
    inputData = scoresExample,
    metadata = metadata1,
    var2predict = "group",
    models = methodsML("svmLinear",
        outcomeClass = "character"
    ),
    Koutter = 2,
    Kinner = 2,
    repeatsCV = 1
)

externalScores <- getScores(dataset2, geneSets = "tmod", method = "Z-score")
realValues <- metadata2$group
names(realValues) <- rownames(metadata2)
predictions <- predictExternal(externalScores, trainedModel,
    realValues = realValues
)

print(predictions)

}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\author{
Iván Ellson, \email{ivan.ellson.l@gmail.com }

Jordi Martorell-Marugán, \email{jordi.martorell@genyo.es}

Daniel Toro-Dominguez, \email{danieltorodominguez@gmail.com}
}
