% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictExternal.R
\name{predictExternal}
\alias{predictExternal}
\title{Apply predictive model to a new dataset}
\usage{
predictExternal(testData, model, orig.values = NULL, positiveClass = NULL)
}
\arguments{
\item{testData}{Must be a numerical matrix containing the same features used
for the model construction in columns, and the samples (new observations) in
the rows}

\item{model}{getML output or a caret-like model object}

\item{orig.values}{Optional, named vector (for numerical variables) or
named factor (for categorical variables) with real values for each sample}

\item{positiveClass}{Optional, positive class to get confusion matrix.
Only needed when orig.values = TRUE and for categorical variables}
}
\value{
A dataframe with prediction for the new observations (if orig.values
is not provided) or a list with the dataframe with predictions and a
dataframe with the performance metrics (if orig.values is provided)
}
\description{
Apply predictive model to a new dataset
}
\references{
Toro-Domínguez, D. et al (2022). \emph{Scoring personalized
molecular portraits identify Systemic Lupus Erythematosus subtypes and
predict individualized drug responses, symptomatology and
disease progression}
 . Briefings in Bioinformatics. 23(5)
}
\author{
Iván Ellson, \email{ivan.ellson.l@gmail.com }

Jordi Martorell-Marugán, \email{jordi.martorell@genyo.es}

Daniel Toro-Dominguez, \email{danieltorodominguez@gmail.com}
}
