% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refData-data.R
\docType{data}
\name{refData}
\alias{refData}
\title{Example of reference gene expression data}
\format{
An object of class \code{"list"} with...
}
\usage{
data(refData)
}
\description{
Data from GSE...
}
\keyword{datasets}
